requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
redis>=4.0.2
celery>=5.1.2
faiss-cpu>=1.7.4
chromadb>=0.4.15
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiohttp>=3.9.1
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2
python-dateutil>=2.8.2
pytz>=2023.3
transformers
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
pydantic-settings>=2.0.0
jose